// src/app/models/masters/exchange-rates-suppliers.ts
export class ExchangeRatesSupplier {
        idExchangeRateSupplier: number = -1;
        idSupplierCompany: number = -1;
        idCurrency: number = -1
        idcompany: number = -1;
        currency: string = "";
        exchangeRate: number = 0;
        createDate: Date;
        createDatestring: string = "";
        UpdatedDate: Date;
        idCreatebyUser: number = -1;
        createbyUser: string = "";
        idUpdatebyuser: number = -1
        updatebyUser: string = "";
        indSupplier = true
}